Rename numba straight guide implementation to OO guide.
