+++ acc/components/optics/oo_guide.py
